print number 3

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,7 +8,7 @@
  */
 void print_number(int n)
 {
-	int l, m, i, j, div;
+	long l, m, i, j, div;
 
 	l = n;
 	m = 0;
